+++ src/render/loader.js
@@ -48,7 +48,8 @@ function load(path, cb) {
 			loaded[pathName] = {
 				source: prerenderedData.content,
 				tags: prerenderedData.list,
-				vars: prerenderedData.vars
+				vars: prerenderedData.vars,
+				literals: prerenderedData.literals
 			};
 			logger.verbose('Pre-rendered:', '[' + pathName + ']', 'from', path);
 			cb();