+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This is a list of manually mantained changes and updates for each version.
 
 ## Added
 
+#### Render added literals
+
 #### Render added custom functions for template variables
 
 ## Changed